:recycle: splitting commands into several files
