+++ src/main/kotlin/com/github/pool_party/pull_party_bot/database/Transactions.kt
@@ -3,10 +3,10 @@ package com.github.pool_party.pull_party_bot.database
 import com.github.pool_party.pull_party_bot.Configuration
 import mu.KotlinLogging
 import org.jetbrains.exposed.sql.Database
-import org.jetbrains.exposed.sql.SchemaUtils
-import org.jetbrains.exposed.sql.StdOutSqlLogger
 import org.jetbrains.exposed.sql.Transaction
 import org.jetbrains.exposed.sql.addLogger
+import org.jetbrains.exposed.sql.SchemaUtils
+import org.jetbrains.exposed.sql.StdOutSqlLogger
 import org.jetbrains.exposed.sql.transactions.transaction
 import java.time.LocalDateTime
