+++ src/main/kotlin/com/github/pool_party/pull_party_bot/database/dao/PartyDao.kt
